Refactoring repositories, demo-admin using utils, creating exceptions and entities

+++ src/infrastructure/db/repositories/UserRepositoryImpl.js
@@ -1,44 +1,68 @@
 const ICommonRepository = require('../../../domain/repositories/ICommonRepository')
 const UserModel = require('../models/user');
+const { DuplicateUserException } = require('../../../../shared/exceptions/DuplicateUserException');
 
 class UserRepositoryImpl extends ICommonRepository {
 
   async getAll() {
-    return await UserModel.findAndCountAll();
+    try {
+      return await UserModel.findAndCountAll();
+    }
+    catch(error){
+      throw new Error('Error getting all users in database:' + error.message)
+    }
   };
 
   async get(numDni) {
-    return await UserModel.findOne({ where: {dni: numDni} });
+    try {
+      return await UserModel.findOne({ where: {dni: numDni} });
+    } 
+    catch (error) {
+      throw new Error('Error getting a user in database:' + error.message)
+    }
   };
 
-  async add() {
-    return await UserModel.create({
-      dni: user.dni,
-      fullName: user.fullName,
-      age: user.age,
-      phone: user.phone,
-      email: user.email,
-      gender: user.gender
-    })
+  async add(user) {
+    try {
+      const existeUsuario = await UserModel.findOne({ where: {dni: user.dni} });
+
+      if(existeUsuario) {
+        throw new DuplicateUserException(); 
+      }
+      return await UserModel.create(user)
+    }
+    catch (error) {
+      throw new Error('Error adding user in database:' + error.message)
+    }
   };
 
-  async edit(numDni, user) {
-    await UserModel.update({
-      fullName: user.fullName,
-      age: user.age,
-      phone: user.phone,
-      email: user.email,
-      gender: user.gender
-    }, 
-    { where: {dni: numDni} });
+  async edit(user) {
+    try {
+      const existeUsuario = await UserModel.findOne({ where: {dni: user.dni} });
 
-    return await get(numDni);
+      if(existeUsuario == null) {
+        throw new Error("User not found")
+      }
+      return await UserModel.update(user, { where: {dni: user.dni} });
+    } 
+    catch (error) {
+      throw new Error("Error editing user in database" + error.message)
+    }
   };
 
   async remove(numDni) {
-    return await UserModel.destroy({ where: {dni: numDni} });
+    try {
+      const existeUsuario = await UserModel.findOne({ where: {dni: numDni} });
+
+      if(existeUsuario == null) {
+        throw new Error("User not found")
+      }
+      return await UserModel.destroy({ where: {dni: numDni} });
+    } 
+    catch (error) {
+      throw new Error("Error deleting user in database" + error.message)
+    }
   };
-  
 }
 
 module.exports = UserRepositoryImpl;
